feat: debug mode option
Added a debug mode that can be enabled by creating a new Equinox client with `equinox.NewClientWithDebug()`, not sure if this implementation of logging is good, only logging requests before and after they are made at the moment.

Moved and added some constants to an api module.

Added go vet and verify to the go workflow.