feat: cache
This commit adds a very simple cache mechanism for GET requests, all successful GET requests are added to a map with its key being the complete URL for the request.

A default Equinox client comes with 120 seconds of TTL for all cached responses. TTL is checked when using cache.Get(), if time.Now() is greater than the expire field inside the CacheItem, the CacheItem is deleted from the cache.

Providing a TTL of 0 inside a custom client disables caching.

There might be unexpected problems coming from this implementation as it is my first time implementing cache in go.

InternalClient.Do() was split in functions for each http Method.