feat: initial rate limit implementation
This commit adds an implementation of rate limiting and a goroutine for the cache, this is my first use of goroutines and it doesn't look any good.

All endpoint methods now sends their endpoint and method name to the internal client, this is used to check and set the rate limit for the method and application, maybe the logs inside the internal client should also log this?

Rate limiting and Cache needs more tests and definitely more work on them.